third_party/blink/web_tests/virtual/synchronous_html_parser/external/wpt/html/semantics/embedded-content/media-elements/track/track-element/track-selection-metadata-expected.txt

This is a testharness.js-based test.
FAIL Multiple 'metadata' tracks with 'default' assert_equals: expected "hidden" but got "disabled"
Harness: the test ran to completion.
